feat(billing): add billing domain, supervisor approval pipeline, and note approval model
Domain models
- Add BillingPeriod and ClaimLine entities (Models/Billing/)
- Add BillingStatus enum (Draft/Submitted/Accepted/Rejected)
- Add PlaceOfService enum with CMS-aligned values
- Add MaineCareId, DiagnosisCode, PlaceOfService to Person
- Add Npi to Agency
- Add Approved/Returned fields to Note (Status values, ApprovedById,
  ApprovedAt, ReturnedById, ReturnedAt, ReturnReason)
- Change Note.Units from int to decimal for fractional unit support

Note lifecycle
- Add Approved and Returned to NoteStatus enum
- Approval pipeline: Draft -> Logged -> Approved/Returned
- ClaimLine snapshot created at approval time (immutable billing record)

Services
- Add IBillingService/BillingService: billing periods, claim line generation,
  period submission
- Add ISupervisorService/SupervisorService: pending note queries, approve, return
- Add GetSuperviseesAsync to IUserService (replaces stale session-cached data)
- Register new services in DI container

Bug fixes
- Fix supervisor dashboard showing zero case managers (login query did not
  eager-load Supervisees; now queried fresh via UserService)
- Fix CaseloadMatrixViewModel null Settings crash on tab reactivation
  (remove premature constructor initialization; let LoadAsync own construction)
- Fix duplicate CaseManagerDashboardViewModel DI registration (remove
  stray AddTransient that overrode AddSingleton)
- Fix namespace reference for Supervisor subfolder in App.xaml DataTemplate
- Rename SupervisorDashboardWindow reference (legacy modal-era naming)

Migrations
- AddBillingAndPersonFields
- AddNoteApprovalFields
- UnitstoDecimal
- Manual SQL fixes for ApprovedById, ReturnedById, and Units column
  conversion due to silently-failed prior migration

Notes
- CaseloadMatrixViewModel folder structure kept flat (existing references)
- New billing code organized in Billing subfolders under Models, Data,
  ViewModels, Views namespaces

+++ Models/Incentive.cs
@@ -21,7 +21,7 @@ namespace Sati.Models
         public int Threshold => DaysScheduled * UnitsPerDay;
 
         public string ExcludedDatesJson { get; set; } = "[]";
-        public decimal Calculate(int loggedUnits)
+        public decimal Calculate(decimal loggedUnits)
         {
             if (loggedUnits < Threshold) return 0;
             if (loggedUnits == Threshold) return BaseIncentive;
